customer_support_email/pyproject.toml
[project]
name = "customer-support-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Humza134", email = "mhumza13420@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.72",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[project.scripts]
invoke = "customer_support_email:main"
stream = "customer_support_email:stream"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
